Change the param back to fp16 and remove some opt
In GPT3_8B model, if we use fp32 param (and autocast),
the chunk moving time will be 14s out of a total 27s,
however, using fp16 param will reduce the moving time
to 6s. The performance loss is too much to neglected and
we have to give up using autocast and turn to apex O2.

+++ patrickstar/core/client.py
@@ -95,7 +95,7 @@ class PatrickStarClient:
         chunk = self.chunk_list.new_chunk()
 
         dummy = torch.nn.Parameter(
-            torch.tensor([], dtype=torch.float), requires_grad=False
+            torch.tensor([], dtype=torch.half), requires_grad=False
         )
         # Add a dummy param to dummy chunk, so that the chunk can be set in HOLD state.
         register_param(dummy, ParamType.CHUNK_BASED, "dummy")
